# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kapital API Documentation",
    "version": "1.0.0",
    "description": "API documentation for Kapital backend services"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/users/me": {
      "get": {
        "summary": "Get current user information",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "phoneNumber": {
                      "type": "string"
                    },
                    "fullName": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "dateOfBirth": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "address1": {
                      "type": "string"
                    },
                    "address2": {
                      "type": "string"
                    },
                    "city": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    },
                    "postalCode": {
                      "type": "string"
                    },
                    "employmentStatus": {
                      "type": "string"
                    },
                    "employerName": {
                      "type": "string"
                    },
                    "monthlyIncome": {
                      "type": "string"
                    },
                    "bankName": {
                      "type": "string"
                    },
                    "accountNumber": {
                      "type": "string"
                    },
                    "isOnboardingComplete": {
                      "type": "boolean"
                    },
                    "onboardingStep": {
                      "type": "number"
                    },
                    "kycStatus": {
                      "type": "string"
                    },
                    "lastLoginAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update current user information",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date"
                  },
                  "address1": {
                    "type": "string"
                  },
                  "address2": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "postalCode": {
                    "type": "string"
                  },
                  "employmentStatus": {
                    "type": "string"
                  },
                  "employerName": {
                    "type": "string"
                  },
                  "monthlyIncome": {
                    "type": "string"
                  },
                  "bankName": {
                    "type": "string"
                  },
                  "accountNumber": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User information updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
...
}